pyproject.toml
[project]
name = "tileserver"
version = "0.1.0"
description = "Minimal OpenSlide tile server for whole-slide images"
requires-python = ">=3.10"
dependencies = [
    "openslide-python>=1.3.1",
    "Pillow>=10.0",
]

[tool.uv]
dev-dependencies = []
